additional tests

+++ contracts/formulas/SushiLPFormula.sol
@@ -15,13 +15,13 @@ contract SushiLPFormula is VotingPowerFormula {
     address public owner;
 
     /// @notice Conversion rate of token to voting power (measured in bips: 10,000 bips = 1%)
-    uint16 public conversionRate;
+    uint32 public conversionRate;
 
     /// @notice Event emitted when the owner of the contract is updated
     event ChangedOwner(address indexed oldOwner, address indexed newOwner);
 
     /// @notice Event emitted when the conversion rate of the contract is changed
-    event ConversionRateChanged(uint16 oldRate, uint16 newRate);
+    event ConversionRateChanged(uint32 oldRate, uint32 newRate);
 
     /// @notice only owner can call function
     modifier onlyOwner {
@@ -34,12 +34,12 @@ contract SushiLPFormula is VotingPowerFormula {
      * @param _owner contract owner
      * @param _cvrRate the conversion rate in bips
      */
-    constructor(address _owner, uint16 _cvrRate) {
+    constructor(address _owner, uint32 _cvrRate) {
         owner = _owner;
         emit ChangedOwner(address(0), owner);
 
         conversionRate = _cvrRate;
-        emit ConversionRateChanged(uint16(0), conversionRate);
+        emit ConversionRateChanged(uint32(0), conversionRate);
     }
 
     /**
@@ -55,7 +55,7 @@ contract SushiLPFormula is VotingPowerFormula {
      * @notice Set conversion rate of contract
      * @param newConversionRate New conversion rate
      */
-    function setConversionRate(uint16 newConversionRate) external onlyOwner {
+    function setConversionRate(uint32 newConversionRate) external onlyOwner {
         emit ConversionRateChanged(conversionRate, newConversionRate);
         conversionRate = newConversionRate;
     }
